lesson_013/requirements.txt
asgiref==3.2.7
astroid==2.3.3
backcall==0.1.0
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
decorator==4.4.2
Django==3.0.5
dnspython==1.16.0
ExifRead==2.1.2
Extractor==0.5
idna==2.9
image==1.5.30
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysql-connector-python==8.0.19
numpy==1.18.2
pandas==1.0.3
parso==0.7.0
pickleshare==0.7.5
Pillow==7.1.1
prompt-toolkit==3.0.5
protobuf==3.11.3
psycopg2==2.8.5
pygame==2.0.0.dev6
Pygments==2.6.1
pylint==2.4.4
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2019.3
requests==2.23.0
simple-draw==2.6.6
six==1.14.0
sqlparse==0.3.1
termcolor==1.1.0
traitlets==4.3.3
urllib3==1.25.9
utils==1.0.1
walk==0.3.5
wcwidth==0.1.9
wrapt==1.12.1
